Move MetricWatcher to prevent circular import in the future
In an upcoming commit, I'll be adding creating SplunkMetricWatchers in a
RollbackDeploymentProcess, but the way things are currently laid out
means that SplunkMetricWatcher will need to import MetricWatcher from
metrics.py - but metrics.py would also need to import from the file that
holds the SplunkMetricWatcher definition, leading to a circular import
